fix:foto ilang pas rename

+++ src/admins/admins.service.ts
@@ -19,20 +19,23 @@ export class AdminsService {
 
   async updateProfile(id: string, photoPath: string | undefined, name: string): Promise<AdminDocument | null> {
     const admin = await this.adminModel.findById(id);
+
     if (!admin) {
       throw new NotFoundException('Admin tidak ditemukan');
     }
 
     const updateData: any = { name };
 
-    if (photoPath) {
+    if (photoPath && photoPath.trim() !== "") { 
       if (admin.photo && admin.photo !== 'puskesmasLogo.png') {
         const oldPath = path.join(process.cwd(), 'public/profiles', admin.photo);
         if (fs.existsSync(oldPath)) {
           fs.unlinkSync(oldPath);
         }
       }
-      updateData.photo = photoPath;
+      updateData.photo = photoPath; 
+    } else {
+      updateData.photo = admin.photo; 
     }
 
     const updatedAdmin = await this.adminModel
@@ -41,7 +44,6 @@ export class AdminsService {
       .exec();
 
     return updatedAdmin;
-  }
-
 }
 
+}
